+++ src/controller/map/simplexml/node/FortificationNode.java
@@ -63,7 +63,7 @@ public class FortificationNode extends AbstractFixtureNode<FortificationEvent> {
 				TownSize.parseTownSize(getProperty(SIZE_PROPERTY)),
 				Integer.parseInt(getProperty(DC_PROPERTY)),
 				hasProperty(NAME_PROPERTY) ? getProperty(NAME_PROPERTY)
-						: "", Long.parseLong(getProperty(ID_PROPERTY)));
+						: "", Integer.parseInt(getProperty(ID_PROPERTY)));
 		if (hasProperty("file")) {
 			fix.setFile(getProperty("file"));
 		}
