+++ src/controller/map/simplexml/node/SimurghNode.java
@@ -27,7 +27,7 @@ public class SimurghNode extends AbstractFixtureNode<Simurgh> {
 	 */
 	@Override
 	public Simurgh produce(final PlayerCollection players, final Warning warner) throws SPFormatException {
-		final Simurgh fix = new Simurgh(Long.parseLong(getProperty("id")));
+		final Simurgh fix = new Simurgh(Integer.parseInt(getProperty("id")));
 		if (hasProperty("file")) {
 			fix.setFile(getProperty("file"));
 		}
