+++ src/model/map/fixtures/Centaur.java
@@ -18,7 +18,7 @@ public class Centaur implements TileFixture, HasImage {
 	 * @param centKind the kind of centaur
 	 * @param idNum the ID number.
 	 */
-	public Centaur(final String centKind, final long idNum) {
+	public Centaur(final String centKind, final int idNum) {
 		kind = centKind;
 		id = idNum;
 	}
@@ -72,7 +72,7 @@ public class Centaur implements TileFixture, HasImage {
 	 */
 	@Override
 	public int hashCode() {
-		return (int) id;
+		return id;
 	}
 	/**
 	 * @param fix a TileFixture to compare to
@@ -85,12 +85,12 @@ public class Centaur implements TileFixture, HasImage {
 	/**
 	 * ID number.
 	 */
-	private final long id; // NOPMD
+	private final int id; // NOPMD
 	/**
 	 * @return a UID for the fixture.
 	 */
 	@Override
-	public long getID() {
+	public int getID() {
 		return id;
 	}
 	/**
